git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/branches/cassinirings@5009 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/qisis/objs/MosaicFileList/MosaicFileListWidget.cpp
@@ -129,7 +129,7 @@ namespace Isis {
       key = key.replace(" ", "_");
       bool visible = !p_tree->isColumnHidden(col);
 
-      output += PvlKeyword(key, toString(visible));
+      output += PvlKeyword(key, toString((int)visible));
       col = (MosaicTreeWidgetItem::TreeColumn)(col + 1);
     }
